* include/bits/istream.tcc (getline): Make sure arguments passed to min/max have the same type. * include/bits/locale_facets.tcc (__verify_grouping): Likewise. * include/bits/stl_deque.h (_Deque_base::_Deque_base): Likewise.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@83522 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libstdc++-v3/include/bits/locale_facets.tcc
@@ -2448,7 +2448,7 @@ namespace std
 		    const string& __grouping_tmp)
   {
     const size_t __n = __grouping_tmp.size() - 1;
-    const size_t __min = std::min(__n, __grouping_size - 1);
+    const size_t __min = std::min(__n, size_t(__grouping_size - 1));
     size_t __i = __n;
     bool __test = true;
